+++ src/js/mixin/parallax.js
@@ -4,7 +4,6 @@ import {
     createEvent,
     css,
     Dimensions,
-    each,
     findIndex,
     isNumber,
     isString,
@@ -65,7 +64,9 @@ export default {
 
     methods: {
         reset() {
-            each(this.getCss(0), (_, prop) => css(this.$el, prop, ''));
+            for (const prop in this.getCss(0)) {
+                css(this.$el, prop, '');
+            }
         },
 
         getCss(percent) {
@@ -79,12 +80,16 @@ export default {
 };
 
 function transformFn(prop, el, stops) {
-    const unit = getUnit(stops) || { x: 'px', y: 'px', rotate: 'deg' }[prop] || '';
+    let unit = getUnit(stops) || { x: 'px', y: 'px', rotate: 'deg' }[prop] || '';
     let transformFn;
 
     if (prop === 'x' || prop === 'y') {
         prop = `translate${ucfirst(prop)}`;
         transformFn = (stop) => toFloat(toFloat(stop).toFixed(unit === 'px' ? 0 : 6));
+    } else if (prop === 'scale') {
+        unit = '';
+        transformFn = (stop) =>
+            getUnit([stop]) ? toPx(stop, 'width', el, true) / el.offsetWidth : stop;
     }
 
     if (stops.length === 1) {